+++ examples/synth/src/synth/oscillator.rs
